+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New verb 'extract' to extract sections of a changelog to a file (#22)
+
 ### Changed
 - Hide exception stacktraces unless verbose logging is enabled (#21)
 - Warn when trying to release version that is older than a previous released version (#23)